requirements.txt
#sds011
#pysds011
sds011lib

pyserial
paho-mqtt

